♻️ Refactor: Move components from Pages to relavent folders
