--- client/src/pages/ArtistPage.tsx
@@ -1,7 +1,0 @@
-import Artist from '../features/artist/components/Artist.tsx';
-
-const ArtistPage = () => {
-  return <Artist />;
-};
-
-export default ArtistPage;
