--- client/src/pages/SearchPage.tsx
@@ -1,11 +1,0 @@
-import Search from '../features/search/Search.tsx';
-
-const SearchPage = () => {
-  return (
-    <div>
-      <Search />
-    </div>
-  );
-};
-
-export default SearchPage;
