--- client/src/pages/helper/NotFound.tsx
@@ -1,5 +1,0 @@
-const NotFound = () => {
-  return <div>404 - not found</div>;
-};
-
-export default NotFound;
